# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AvailabilityView": {
        "properties": {
          "available_quantity": {
            "minimum": 0.0,
            "title": "Available Quantity",
            "type": "integer"
          },
          "in_stock": {
            "title": "In Stock",
            "type": "boolean"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "sale_status": {
            "enum": [
              "draft",
              "active",
              "inactive"
            ],
            "title": "Sale Status",
            "type": "string"
          }
        },
        "required": [
          "sale_status",
          "available_quantity",
          "in_stock"
        ],
        "title": "AvailabilityView",
        "type": "object"
      },
...
      "CatalogCategoryView": {
        "additionalProperties": false,
        "properties": {
          "available_product_count": {
            "minimum": 0.0,
            "title": "Available Product Count",
            "type": "integer"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "product_count": {
            "minimum": 0.0,
            "title": "Product Count",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "display_name",
          "product_count",
          "available_product_count"
        ],
        "title": "CatalogCategoryView",
        "type": "object"
      },
      "CatalogErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "unsupported_category",
              "catalog_not_found",
              "catalog_unavailable",
              "catalog_data_invalid"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "CatalogErrorResponse",
        "type": "object"
      },
      "CatalogProductDetail": {
        "additionalProperties": false,
        "properties": {
          "brand": {
            "title": "Brand",
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/CatalogCategoryView"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "product_code": {
            "title": "Product Code",
            "type": "string"
          },
          "product_id": {
            "format": "uuid",
            "title": "Product Id",
            "type": "string"
          },
          "skus": {
            "items": {
              "$ref": "#/components/schemas/CatalogSkuView"
            },
            "minItems": 1,
            "title": "Skus",
            "type": "array"
          },
          "specifications": {
            "items": {
              "$ref": "#/components/schemas/CatalogSpecificationView"
            },
            "title": "Specifications",
            "type": "array"
          }
        },
        "required": [
          "product_id",
          "product_code",
          "brand",
          "name",
          "category",
          "skus"
        ],
        "title": "CatalogProductDetail",
        "type": "object"
      },
      "CatalogProductListResponse": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/CatalogCategoryView"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogProductSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "category",
          "total",
          "limit",
          "offset"
        ],
        "title": "CatalogProductListResponse",
        "type": "object"
      },
      "CatalogProductSummary": {
        "additionalProperties": false,
        "properties": {
          "brand": {
            "title": "Brand",
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/CatalogCategoryView"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "product_code": {
            "title": "Product Code",
            "type": "string"
          },
          "product_id": {
            "format": "uuid",
            "title": "Product Id",
            "type": "string"
          },
          "skus": {
            "items": {
              "$ref": "#/components/schemas/CatalogSkuView"
            },
            "minItems": 1,
            "title": "Skus",
            "type": "array"
          },
          "specifications": {
            "items": {
              "$ref": "#/components/schemas/CatalogSpecificationView"
            },
            "title": "Specifications",
            "type": "array"
          }
        },
        "required": [
          "product_id",
          "product_code",
          "brand",
          "name",
          "category",
          "skus"
        ],
        "title": "CatalogProductSummary",
        "type": "object"
      },
      "CatalogSkuView": {
        "additionalProperties": false,
        "properties": {
          "availability": {
            "$ref": "#/components/schemas/AvailabilityView"
          },
          "money": {
            "$ref": "#/components/schemas/Money"
          },
          "sku_code": {
            "title": "Sku Code",
            "type": "string"
          },
          "sku_id": {
            "format": "uuid",
            "title": "Sku Id",
            "type": "string"
          },
          "sku_name": {
            "title": "Sku Name",
            "type": "string"
          },
          "variant_specifications": {
            "items": {
              "$ref": "#/components/schemas/CatalogSpecificationView"
            },
            "title": "Variant Specifications",
            "type": "array"
          }
        },
        "required": [
          "sku_id",
          "sku_code",
          "sku_name",
          "money",
          "availability"
        ],
        "title": "CatalogSkuView",
        "type": "object"
      },
      "CatalogSpecificationView": {
        "additionalProperties": false,
        "description": "Registry-labelled Catalog fact safe for generic browser rendering.",
        "properties": {
          "comparable": {
            "default": false,
            "title": "Comparable",
            "type": "boolean"
          },
          "display_order": {
            "default": 0,
            "title": "Display Order",
            "type": "integer"
          },
          "format_hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format Hint"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              },
              {
                "type": "boolean"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Value"
          },
          "value_type": {
            "enum": [
              "number",
              "string",
              "enum",
              "boolean",
              "string_list"
            ],
            "title": "Value Type",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "value",
          "value_type"
        ],
        "title": "CatalogSpecificationView",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Money": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "currency"
        ],
        "title": "Money",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "FastAPI backend skeleton for the RetailPilot / ShopMind project.",
    "title": "RetailPilot Backend API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/catalog/products": {
      "get": {
        "operationId": "list_catalog_products_api_catalog_products_get",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Category",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogProductListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Catalog Products",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/catalog/products/{product_code}": {
      "get": {
        "operationId": "get_catalog_product_api_catalog_products__product_code__get",
        "parameters": [
          {
            "in": "path",
            "name": "product_code",
            "required": true,
            "schema": {
              "title": "Product Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogProductDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Catalog Product",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}